Extract TUI I/O into client.go; model delegates through Client
All disk reads, git subprocess calls, and event log tailing now live in
client.go. The Model holds a *Client and loadData() is a one-liner that
calls client.load(m). Pure logic (annotateActivity, sortByActivity,
mergeBranches, filterSidecars, capEvents) stays in model.go.
